extracted future builders to separate widgets to clean up dashboard.dart
