+++ lib/utils/date_formatter.dart
@@ -1,0 +1,4 @@
+import 'package:intl/intl.dart';
+
+final currentDayFormatter = DateFormat('h:mm a');
+final recordDateFormatter = DateFormat.yMd().add_jm();
